WIP: restructuring fixpnt test suites
